+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2024-11-08
+
+### Fixed
+- Modified Topsort's app name to get settings from VTEX
+
 ## [1.0.1] - 2024-10-28
 
 ### Added